backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
python-dotenv
flask-jwt-extended
flask-bcrypt